Rename accordion.js to accordion-anchors.js (#2296)
