+++ js/index.umd.js
@@ -19,7 +19,7 @@ import Toast from '../node_modules/bootstrap/js/src/toast.js'
 import Tooltip from '../node_modules/bootstrap/js/src/tooltip.js'
 import fixModalAriaHidden from './src/modal.js'
 import photoGalleryGridSlideToImage from './src/photogallery.js'
-import enableAccordionAnchors from './src/accordion.js'
+import enableAccordionAnchors from './src/accordion-anchors.js'
 import enableAzNavbar from './src/navbar.js'
 import enableNavbarAzFullscreen from './src/navbar-az-fullscreen.js'
 import enableNavbarAzFullscreenMobileNav from './src/navbar-az-fullscreen-mobile-nav.js'
